#8
Before appending "email:", we check if the nodeName already contains it
If it does, we don't append it again
If it doesn't, we append it as before

This will prevent the double "email:" issue while maintaining the expected format for account IDs.

Will need tested made multiple accounts - seems to be fine so far.